+++ web-app/app/(discovery)/(home)/interest-selection/page.tsx
@@ -1,0 +1,5 @@
+import { InterestSelection } from "./interest-selection";
+
+export default function InterestSelectionPage() {
+    return <InterestSelection />;
+}
